review tests refactoring
