+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/planner/AggregatePlannerTest.java
@@ -17,21 +17,15 @@
 
 package org.apache.ignite.internal.processors.query.calcite.planner;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import org.apache.calcite.plan.RelOptPlanner;
 import org.apache.calcite.plan.RelOptUtil;
 import org.apache.calcite.rel.RelCollations;
-import org.apache.calcite.rel.type.RelDataTypeFactory;
 import org.apache.calcite.sql.SqlExplainLevel;
 import org.apache.calcite.sql.fun.SqlAvgAggFunction;
-import org.apache.calcite.sql.fun.SqlCountAggFunction;
 import org.apache.calcite.util.ImmutableIntList;
-import org.apache.ignite.internal.processors.query.calcite.metadata.ColocationGroup;
-import org.apache.ignite.internal.processors.query.calcite.prepare.PlanningContext;
 import org.apache.ignite.internal.processors.query.calcite.rel.IgniteAggregate;
 import org.apache.ignite.internal.processors.query.calcite.rel.IgniteAggregateBase;
 import org.apache.ignite.internal.processors.query.calcite.rel.IgniteHashAggregate;
@@ -45,14 +39,10 @@ import org.apache.ignite.internal.processors.query.calcite.rel.IgniteRel;
 import org.apache.ignite.internal.processors.query.calcite.rel.IgniteSort;
 import org.apache.ignite.internal.processors.query.calcite.rel.IgniteSortAggregate;
 import org.apache.ignite.internal.processors.query.calcite.schema.IgniteSchema;
-import org.apache.ignite.internal.processors.query.calcite.trait.IgniteDistribution;
-import org.apache.ignite.internal.processors.query.calcite.trait.IgniteDistributions;
 import org.apache.ignite.internal.processors.query.calcite.type.IgniteTypeFactory;
 import org.apache.ignite.internal.processors.query.calcite.type.IgniteTypeSystem;
 import org.apache.ignite.internal.util.typedef.F;
-import org.apache.ignite.testframework.GridTestUtils;
 import org.hamcrest.core.IsInstanceOf;
-import org.jetbrains.annotations.NotNull;
 import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -64,7 +54,7 @@ import org.junit.runners.Parameterized;
  */
 @SuppressWarnings({"TypeMayBeWeakened"})
 @RunWith(Parameterized.class)
-public class AggregatePlannerTest extends AbstractPlannerTest {
+public class AggregatePlannerTest extends AbstractAggregatePlannerTest {
     /** Algorithm. */
     @Parameterized.Parameter
     public AggregateAlgorithm algo;
@@ -227,121 +217,6 @@ public class AggregatePlannerTest extends AbstractPlannerTest {
 
         if (algo == AggregateAlgorithm.SORT)
             assertNotNull(findFirstNode(phys, byClass(IgniteIndexScan.class)));
-    }
-
-    /**
-     *
-     */
-    @Test
-    public void notApplicableForSortAggregate() {
-        if (algo == AggregateAlgorithm.SORT)
-            return;
-
-        TestTable tbl = createAffinityTable().addIndex(RelCollations.of(ImmutableIntList.of(1, 2)), "val0_val1");
-
-        IgniteSchema publicSchema = new IgniteSchema("PUBLIC");
-
-        publicSchema.addTable("TEST", tbl);
-
-        String sqlMin = "SELECT MIN(val0) FROM test";
-
-        GridTestUtils.assertThrows(log,
-            () -> physicalPlan(
-                sqlMin,
-                publicSchema,
-                "HashAggregateConverterRule"
-            ),
-            RelOptPlanner.CannotPlanException.class,
-            "There are not enough rules to produce a node with desired properties"
-        );
-    }
-
-    /**
-     *
-     */
-    @Test
-    public void noGroupByAggregate() throws Exception {
-        if (algo == AggregateAlgorithm.SORT)
-            return;
-
-        TestTable tbl = createAffinityTable().addIndex(RelCollations.of(ImmutableIntList.of(1, 2)), "val0_val1");
-
-        IgniteSchema publicSchema = new IgniteSchema("PUBLIC");
-
-        publicSchema.addTable("TEST", tbl);
-
-        String sqlCount = "SELECT COUNT(*) FROM test";
-
-        IgniteRel phys = physicalPlan(
-            sqlCount,
-            publicSchema,
-            algo.ruleToDisable
-        );
-
-        checkSplitAndSerialization(phys, publicSchema);
-
-        IgniteAggregate mapAgg = findFirstNode(phys, byClass(algo.map));
-        IgniteReduceAggregateBase rdcAgg = findFirstNode(phys, byClass(algo.reduce));
-
-        assertNotNull("Invalid plan\n" + RelOptUtil.toString(phys), rdcAgg);
-        assertNotNull("Invalid plan\n" + RelOptUtil.toString(phys), mapAgg);
-
-        Assert.assertThat(
-            "Invalid plan\n" + RelOptUtil.toString(phys),
-            F.first(rdcAgg.aggregateCalls()).getAggregation(),
-            IsInstanceOf.instanceOf(SqlCountAggFunction.class));
-
-        Assert.assertThat(
-            "Invalid plan\n" + RelOptUtil.toString(phys),
-            F.first(mapAgg.getAggCallList()).getAggregation(),
-            IsInstanceOf.instanceOf(SqlCountAggFunction.class));
-    }
-
-    @NotNull private TestTable createBroadcastTable() {
-        IgniteTypeFactory f = new IgniteTypeFactory(IgniteTypeSystem.INSTANCE);
-
-        TestTable tbl = new TestTable(
-            new RelDataTypeFactory.Builder(f)
-                .add("ID", f.createJavaType(Integer.class))
-                .add("VAL0", f.createJavaType(Integer.class))
-                .add("VAL1", f.createJavaType(Integer.class))
-                .add("GRP0", f.createJavaType(Integer.class))
-                .add("GRP1", f.createJavaType(Integer.class))
-                .build()) {
-
-            @Override public IgniteDistribution distribution() {
-                return IgniteDistributions.broadcast();
-            }
-        };
-        return tbl;
-    }
-
-    @NotNull private AbstractPlannerTest.TestTable createAffinityTable() {
-        IgniteTypeFactory f = new IgniteTypeFactory(IgniteTypeSystem.INSTANCE);
-
-        return new TestTable(
-            new RelDataTypeFactory.Builder(f)
-                .add("ID", f.createJavaType(Integer.class))
-                .add("VAL0", f.createJavaType(Integer.class))
-                .add("VAL1", f.createJavaType(Integer.class))
-                .add("GRP0", f.createJavaType(Integer.class))
-                .add("GRP1", f.createJavaType(Integer.class))
-                .build()) {
-
-            @Override public ColocationGroup colocationGroup(PlanningContext ctx) {
-                return ColocationGroup.forAssignments(Arrays.asList(
-                    select(nodes, 0, 1),
-                    select(nodes, 1, 2),
-                    select(nodes, 2, 0),
-                    select(nodes, 0, 1),
-                    select(nodes, 1, 2)
-                ));
-            }
-
-            @Override public IgniteDistribution distribution() {
-                return IgniteDistributions.affinity(0, "test", "hash");
-            }
-        };
     }
 
     /** */
